requirements-base.txt
# --- Shared Legal Core ---
fastapi
websockets
pydantic
python-jose[cryptography]
python-multipart
boto3
aioboto3>=13.0.0
python-dotenv
sqlalchemy
asyncpg
redis
alembic
pydantic-settings
passlib[bcrypt]
bcrypt
greenlet
apscheduler
google-auth
requests
httpx
qdrant-client
fpdf2
